Add files via upload
 creating the "FilterExample" dynamic web page project! In your package, you have built two servlets named "FirstServlet" and "MyFilter."

The "FirstServlet" handles specific functionalities and interactions with the web page, while the "MyFilter" serves as a filter to intercept and process requests and responses before they reach the "FirstServlet."

The WebContent directory contains the "index.html" file, which is the main web page that users will interact with. Additionally, you have a web.xml file that stores mappings and configurations for the project.

Furthermore, you have connected your web application to a MySQL database using the MySQL Connector, which allows you to perform database operations such as storing and retrieving data.

Overall, "FilterExample" is a dynamic web page project that demonstrates the use of servlets, filters, and database connectivity to create a functional and interactive web application.